requirements.txt
fastapi
uvicorn
PyPDF2
langchain
langchain-openai
langchain-community
langchain-huggingface
faiss-cpu
python-dotenv
